+++ siosocks/io/trio.py
@@ -2,6 +2,7 @@ import logging
 
 import trio
 
+from ..exceptions import SocksException
 from ..interface import AbstractSocksIO, async_engine
 from ..protocol import SocksServer, SocksClient, DEFAULT_ENCODING
 from .const import DEFAULT_BLOCK_SIZE
@@ -82,8 +83,8 @@ async def open_tcp_stream(host, port, *, socks_host=None, socks_port=None,
     socks_enabled = all(socks_required)
     socks_disabled = not any(socks_required)
     if socks_enabled == socks_disabled:
-        raise ValueError("Partly passed socks required arguments: "
-                         "socks_host = {!r}, socks_port = {!r}, socks_version = {!r}".format(*socks_required))
+        raise SocksException("Partly passed socks required arguments: "
+                             "socks_host = {!r}, socks_port = {!r}, socks_version = {!r}".format(*socks_required))
     if socks_enabled:
         stream = await trio.open_tcp_stream(socks_host, socks_port, **open_tcp_stream_extras)
         protocol = SocksClient(host, port, socks_version, username=username, password=password, encoding=encoding,