+++ tests/test_asyncio.py
@@ -3,6 +3,7 @@ import asyncio
 import pytest
 
 from siosocks.io.asyncio import socks_server_handler, open_connection
+from siosocks.exceptions import SocksException
 
 
 HOST = "127.0.0.1"
@@ -52,7 +53,14 @@ async def test_connection_socks_success(endpoint_port, socks_server_port):
 
 
 @pytest.mark.asyncio
-async def test_connection_partly_passed_error():
-    with pytest.raises(ValueError):
+async def test_connection_socks_failed(socks_server_port, unused_tcp_port):
+    with pytest.raises(SocksException):
+        await open_connection(HOST, unused_tcp_port,
+                              socks_host=HOST, socks_port=socks_server_port, socks_version=4)
+
+
+@pytest.mark.asyncio
+async def test_connection_partly_passed_error(endpoint_port, socks_server_port):
+    with pytest.raises(SocksException):
         await open_connection(HOST, endpoint_port,
                               socks_host=HOST, socks_port=socks_server_port)