+++ history.md
@@ -1,2 +1,5 @@
+# x.x.x (xxxx-xx-xx)
+- add unified `SocksException`
+
 # 0.0.2 (2019-08-02)
-* initial release
+- initial release